+++ xinitrc.d/devilspie2.sh
@@ -2,15 +2,13 @@
 
 ## Start devilspie2 in the background as a daemon
 
-test -f "${XDG_RUNTIME_DIR}/devilspie2.pid" && \
-test -r "${XDG_RUNTIME_DIR}/devilspie2.pid" && {
-  kill -15 "$(cat "${XDG_RUNTIME_DIR}/devilspie2.pid")"
-  rm "${XDG_RUNTIME_DIR}/devilspie2.pid"
-} 2>/dev/null
+if command -v compton 1>/dev/null; then
+  if test -f "${XDG_RUNTIME_DIR}/devilspie2.pid"; then
+    if test -r "${XDG_RUNTIME_DIR}/devilspie2.pid"; then
+      kill -15 "$(cat "${XDG_RUNTIME_DIR}/devilspie2.pid")"
+    fi
+    rm "${XDG_RUNTIME_DIR}/devilspie2.pid"
+  fi
 
-devilspie2 & printf '%d\n' "$!" >| "${XDG_RUNTIME_DIR}/devilspie2.pid"
-
-#pkill -15 devilspie2
-#pgrep devilspie2 1>/dev/null && pkill -9 devilspie2
-#
-#devilspie2 &
+  devilspie2 & echo "$!" 1>| "${XDG_RUNTIME_DIR}/devilspie2.pid"
+fi
